+++ packages/app/downloads.ts
@@ -7,6 +7,7 @@ import type { DownloadItem } from "@meru/shared/types";
 import { type BrowserWindow, shell } from "electron";
 import electronDl from "electron-dl";
 import { config } from "@/config";
+import { main } from "@/main";
 import { createNotification } from "@/notifications";
 import { fileExists } from "./lib/fs";
 import { Popup } from "./lib/popup";
@@ -87,17 +88,21 @@ class Downloads {
           const shouldOpenFile =
             config.get("notifications.onClickDownloadCompleted") === "openFile";
 
+          const openDownload = () => {
+            if (shouldOpenFile) {
+              shell.openPath(filePath);
+            } else {
+              shell.showItemInFolder(filePath);
+            }
+          };
+
           createNotification({
             title: `Downloaded ${fileName}`,
             body: shouldOpenFile
               ? "Click to open the file."
               : `Click to show the file in ${FILE_MANAGER_NAME}.`,
             click: () => {
-              if (shouldOpenFile) {
-                shell.openPath(filePath);
-              } else {
-                shell.showItemInFolder(filePath);
-              }
+              main.runAfterNotificationClick(openDownload);
             },
           });
         }